I believe I have context logic now on ChatScreen.
